+++ src/main/java/net/jodah/failsafe/ExecutionContext.java
@@ -16,6 +16,7 @@
 package net.jodah.failsafe;
 
 import java.time.Duration;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Contextual execution information.
@@ -25,7 +26,11 @@ import java.time.Duration;
 public class ExecutionContext {
   private final Duration startTime;
   /** Number of execution attempts */
-  volatile int attempts;
+  AtomicInteger attempts = new AtomicInteger();
+
+  // Internally mutable state
+  volatile Object lastResult;
+  volatile Throwable lastFailure;
 
   ExecutionContext(Duration startTime) {
     this.startTime = startTime;
@@ -34,6 +39,8 @@ public class ExecutionContext {
   private ExecutionContext(ExecutionContext context) {
     this.startTime = context.startTime;
     this.attempts = context.attempts;
+    this.lastResult = context.lastResult;
+    this.lastFailure = context.lastFailure;
   }
 
   /**
@@ -47,7 +54,31 @@ public class ExecutionContext {
    * Gets the number of execution attempts so far.
    */
   public int getAttemptCount() {
-    return attempts;
+    return attempts.get();
+  }
+
+  /**
+   * Returns the last failure that was recorded.
+   */
+  @SuppressWarnings("unchecked")
+  public <T extends Throwable> T getLastFailure() {
+    return (T) lastFailure;
+  }
+
+  /**
+   * Returns the last result that was recorded.
+   */
+  @SuppressWarnings("unchecked")
+  public <T> T getLastResult() {
+    return (T) lastResult;
+  }
+
+  /**
+   * Returns the last result that was recorded else the {@code defaultValue}.
+   */
+  @SuppressWarnings("unchecked")
+  public <T> T getLastResult(T defaultValue) {
+    return lastResult != null ? (T) lastResult : defaultValue;
   }
 
   /**